Revert "entity,repository JPA 수정"
This reverts commit 5c83ba614c47dce6a606a94a64869d0c35d2216e.

+++ src/main/java/antigravity/model/request/ProductInfoRequest.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class ProductInfoRequest {
-    private Long productId;
-    private long[] couponIds;
+    private int productId;
+    private int[] couponIds;
 }
